MH-6336 -- Job versions are now longs, so they don't overflow easily
git-svn-id: https://opencast.jira.com/svn/MH/trunk@9580 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ docs/scripts/ddl/postgres84.sql
@@ -101,7 +101,7 @@ CREATE TABLE job (
     payload text,
     datestarted timestamp without time zone,
     runtime bigint,
-    instance_version integer,
+    instance_version bigint,
     datecompleted timestamp without time zone,
     operation character varying(255),
     datecreated timestamp without time zone,
